refa: 상속을 통한 중복되는 코드 제거
